requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
openai>=1.0.0
pdf2image>=1.16.0
pillow>=9.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
imageio[avif]
numpy>=1.21.0
python-dotenv>=1.1.1
openpyxl>=3.0.10